+++ src/index.js
@@ -45,3 +45,4 @@ require('chartjs-plugin-zoom');
 window.Popper = require('@popperjs/core');
 window.iro = require('@jaames/iro').default;
 window.ICAL = require('ical.js');
+window.workerTimers = require('worker-timers');